refactor(pos): extract CustomerPicker and add customer step to bill split
Reuse search-and-create customer selection in BillSplitDialog with an
optional different customer for the new split child invoice.
